Remove tesseract from repository
Closes #13 for now.

The tesseract code in the repository appeared to only be partially
implemented. Specifically, tesseract was used in loading the test data,
but not in loading the training data--resulting in a ValueError that
crashed the `extract_tables` utility when a scanned PDF was provided.

Eventually this will be replaced using the vision code from Ines, so for
simplicity's sake, we are removing tesseract from the repository for
now, and throwing exceptions where necessary.