docs(README): clarify that pdftotree is unmaintained
`pdftotree` is not only not actively developed, it is not even passively
maintained. It should be considered completely experimntal. That is, we
want to share it with the community, but do not intend to meet anyone's
particular use case.

While pull requests and issues are welcome, there should be no
expectation that code will be reviewed/merged, or that issues will be
addressed.